RE #28355 Add calibration info to status bar
The status bar at the bottom of the GUI is now populated with the
currently loaded calibration.
